set entry point as app.py
