+++ src/app.py
@@ -1,0 +1,5 @@
+from .factory import create_app
+
+app=create_app()
+
+app.run()
