+++ src/factory.py
@@ -1,0 +1,39 @@
+import os
+
+from flask import Flask
+
+def create_app(test_config=None):
+    # load environment variables
+
+    # create and configure the app
+    app = Flask(__name__, instance_relative_config=True)
+    app.config.from_mapping(
+        SECRET_KEY=os.getenv('SECRET_KEY'),
+        DATABASE=os.path.join(app.instance_path, 'urls.sqlite'),
+    )
+
+    if test_config is None:
+        # load the instance config, if it exists, when not testing
+        app.config.from_pyfile('config.py', silent=True)
+    else:
+        # load the test config if passed in
+        app.config.from_mapping(test_config)
+
+    # create the instance folder
+    try:
+        os.makedirs(app.instance_path)
+    except OSError:
+        pass
+
+    #database
+    from . import db
+
+    db.init_app(app)
+
+    # register blue print
+    from . import link_checker
+
+    app.register_blueprint(link_checker.bp)
+
+    #return the application instance
+    return app
