+++ src/schema.sql
@@ -3,5 +3,6 @@ DROP TABLE IF EXISTS urls;
 CREATE TABLE urls(
     status TEXT NOT NULL,
     message TEXT NOT NULL,
-    link TEXT NOT NULL
+    link TEXT NOT NULL,
+    timestamp TIMESTAMP NOT NULL
 )
